requirements.txt
sentence-transformers
openai
torch
google-generativeai
dotenv
ollama pull mistral
pymupdf
langchain
pinecone-client
InstructorEmbedding
faiss-cpu
fastapi
uvicorn